Add motion plugin with pointer-lock controls
