+++ client/crates/engine/Cargo.toml
@@ -7,8 +7,6 @@ edition = "2024"
 bevy = { version = "0.12", default-features = false, features = ["bevy_render", "bevy_core_pipeline", "bevy_pbr", "bevy_text"] }
 platform-api = { path = "../../../crates/platform-api" }
 bevy_rapier3d = { version = "0.24", default-features = false, features = ["dim3"] }
-serde = { version = "1.0", features = ["derive"] }
-toml = "0.8"
 serde_json = "1.0"
 futures-lite = "2.3"
 thiserror = "1"
