+++ crates/platform-api/Cargo.toml
@@ -15,6 +15,8 @@ bevy = { version = "0.12", default-features = false, features = [
 ] }
 bitflags = "2"
 anyhow = "1"
+serde = { version = "1.0", features = ["derive"] }
+toml = "0.8"
 
 [features]
 default = []
